feat: move story view states to viewmodel
