+++ Moco/View/Components/PopUpComponent.swift
@@ -198,20 +198,19 @@ struct PopUpComponentView: View {
                     switch type {
                     case .base:
                         Image(isLarge ?
-                              "Components/popup-base-lg" :
-                                "Components/popup-base")
-                        .resizable()
-                        .scaledToFill()
+                            "Components/popup-base-lg" :
+                            "Components/popup-base")
+                            .resizable()
+                            .scaledToFill()
                     case .danger:
                         Image(isLarge ?
-                              "Components/popup-base-lg" :
-                                "Components/popup-danger")
-                        .resizable()
-                        .scaledToFill()
+                            "Components/popup-base-lg" :
+                            "Components/popup-danger")
+                            .resizable()
+                            .scaledToFill()
                     }
                 }
                 .frame(maxWidth: width, maxHeight: height)
-
             }
             .offset(x: 0, y: offset)
             .onAppear {
